Implement general API to get and update entities
Every phoenix entity has now Key attributes to identify them.

The new class SelectEntity provide an interface to select specific
entities of the entity type. Create them and fill them with attribute
names from the class.
They are used for a more general get and delete interface.
Using them without any attribute set is the same as a getAll command for
the entity. Easier use is accomplished with SelectAllEntity class.

Same goes with the UpdateEntity class, but this class stores the new
object. The old object will replace all attributes with the new ones.

This only goes for simple attributes and not phoenixentity attributes.