Remove redundant JSON wrapping in SelectEntity
All object used as keys in SelectEntity were stored in a Map<String,
String>. The value of the entry was a JSON formatted representation of the
attribute generated by JSON mapper.
But all hash maps are translated into json and all their entries too.
Storing the string instead the object is redundant and was removed.

Also deprecated the getter method, no class is needed anymore.