Add PhoenixLectureGroup entity to library
Also rewrote the complete ConnectionWithEntity system using a more general
for n:m relations with attributes.

Reference implementation is the rewrote API for PhoenixTaskSheet and the
new PhoenixLectureGroupTaskSheet.

A ConnectionEntity hold only references on the entities to connect and
can have more entity types than two. This is not very type safe, so there
are helper classes in the de.phoenix.rs.entity.connection package.